+++ backend/test/data/usecases/user/user-authentication.spec.ts
@@ -1,6 +1,6 @@
 import { HashComparer } from '@/data/protocols/criptography/hash-compare'
 import { DbLoadAccountByRegistration } from '@/data/protocols/db/db-load-account-by-registration'
-import { AuthenticationData } from '@/data/usecases/authentication/auth-authentication'
+import { AuthenticationData } from '@/data/usecases/user/auth-authentication'
 import { User, UserPermission } from '@/domain/model/user'
 import { AuthenticationModel } from '@/domain/usecase/user/user-authentication'
 import Encrypter from '@/data/protocols/criptography/encrypter'
